(GH1994) breaks data out into separate files
