+++ test/tree-tests/data/callCount.js
@@ -1,0 +1,13 @@
+define(function callCountModule () {
+	return {
+		data: [
+			{
+				'name': 'Convex and Concave',
+				'type': 'item',
+				'attr': {
+					'id': 'item4'
+				}
+			}
+		]
+	};
+});
