+++ test/tree-tests/data/guid.js
@@ -1,0 +1,11 @@
+define(function guidModule () {
+	return function guid () {
+		var s4 = function s4 () {
+			return Math.floor((1 + Math.random()) * 0x10000)
+				.toString(16)
+				.substring(1);
+		};
+
+		return s4() + s4() + '-' + s4() + '-' + s4() + '-' + s4() + '-' + s4() + s4() + s4();
+	};
+});
