+++ test/tree-tests/data/textData.js
@@ -1,0 +1,13 @@
+define(function textDataModule () {
+	return {
+		data: [
+			{
+				text: 'node text',
+				type: 'folder',
+				attr: {
+					id: 'folder1'
+				}
+			}
+		]
+	};
+});
