+++ test/tree-tests/helpers.js
@@ -13,112 +13,9 @@ define(function treeDataFactory (require) {
 		NUM_OVERFLOWS: 1
 	};
 
-	var guid = function guid () {
-		var s4 = function s4 () {
-			return Math.floor((1 + Math.random()) * 0x10000)
-				.toString(16)
-				.substring(1);
-		};
-
-		return s4() + s4() + '-' + s4() + '-' + s4() + '-' + s4() + '-' + s4() + s4() + s4();
-	};
-
-	var callCountData = {
-		data: [
-			{
-				'name': 'Convex and Concave',
-				'type': 'item',
-				'attr': {
-					'id': 'item4'
-				}
-			}
-		]
-	};
-
-	var textData = {
-		data: [
-			{
-				text: 'node text',
-				type: 'folder',
-				attr: {
-					id: 'folder1'
-				}
-			}
-		]
-	};
-
-	var treeData = function generateTreeData () {
-		return {
-			data: [
-				{
-					name: 'Ascending and Descending',
-					type: 'folder',
-					attr: {
-						id: 'folder' + guid()
-					}
-				},
-				{
-					name: 'Sky and Water I (with custom icon)',
-					type: 'item',
-					attr: {
-						id: 'folder' + guid(),
-						'data-icon': 'glyphicon glyphicon-file'
-					}
-				},
-				{
-					name: 'Drawing Hands',
-					type: 'folder',
-					attr: {
-						id: 'folder' + guid(),
-						'data-children': false
-					}
-				},
-				{
-					name: 'Waterfall',
-					type: 'item',
-					attr: {
-						id: 'item2'
-					}
-				},
-				{
-					name: 'Belvedere',
-					type: 'folder',
-					attr: {
-						id: 'folder' + guid()
-					}
-				},
-				{
-					name: 'Relativity (with custom icon)',
-					type: 'item',
-					attr: {
-						id: 'item3',
-						'data-icon': 'glyphicon glyphicon-picture'
-					}
-				},
-				{
-					name: 'House of Stairs',
-					type: 'folder',
-					attr: {
-						id: 'folder' + guid()
-					}
-				},
-				{
-					name: 'Convex and Concave',
-					type: 'item',
-					attr: {
-						id: 'item4'
-					}
-				},
-				{
-					name: 'Load More',
-					type: 'overflow',
-					attr: {
-						id: 'overflow1'
-					}
-				}
-			]
-		};
-	};
+	var callCountData = require('./data/callCount');
+	var textData = require('./data/textData');
+	var treeData = require('./data/treeData');
 
 	var setup = function setup () {
 		var callLimit = 50;
